SPlit ssh tests.
